(refactor): uplift logic to game class
