+++ react/tic-tac-toe/src/app/screens/Game/components/Board/index.js
@@ -5,64 +5,17 @@ import Square from '../Square';
 import styles from './styles.module.scss';
 
 class Board extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      squares: Array(9).fill(null),
-      player: 'X',
-    };
-  }
 
-  changeTurn = () => this.state.player === 'X' ? 'O' : 'X';
-
-  handleClick(i) {
-    const squares = this.state.squares.slice();
-    if (this.calculateWinner(squares) || squares[i]) return;
-
-    squares[i] = this.state.player;
-    this.setState({squares : squares, player: this.changeTurn() });
-  }
-
-  renderSquare(i) {
-    return <Square value={this.state.squares[i]} onClick={ ()=> this.handleClick(i)} />;
-  }
-
-  calculateWinner(squares) {
-    const lines = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6],
-    ];
-    for (let i = 0; i < lines.length; i++) {
-      const [a, b, c] = lines[i];
-      if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-        return squares[a];
-      }
-    }
-    return null;
+  renderSquare = i => {
+        return <
+                Square value={this.props.squares[i]}
+                onClick={ ()=> this.props.onClick(i)}
+              />;
   }
 
   render() {
-    const winner = this.calculateWinner(this.state.squares);
-    let status = this.state.player + `'s turn`;
-
-    winner ? status = winner : status = this.state.player + `'s turn`;
-
-    if (winner) {
-      status = 'Winner ' + winner;
-    }
-    else {
-
-    }
-
     return (
       <div>
-        <div className={styles.status}>{status}</div>
         <div className={styles.boardRow}>
           {this.renderSquare(0)}
           {this.renderSquare(1)}
